refactor: move DeviceId to its own module
